+++ src/main/kotlin/forms/SickDayForm.kt
@@ -19,8 +19,8 @@ data class SickDayForm(
     @JsonSerialize(using = LocalDateSerializer::class)
     override val to: LocalDate,
 
-    override val hours: Int,
-    override val days: List<Int>,
+    override val hours: Double,
+    override val days: List<Double>,
     override val status: Status = Status.REQUESTED,
 
     val description: String? = null,
