+++ src/main/kotlin/interfaces/Hours.kt
@@ -1,6 +1,6 @@
 package community.flock.eco.workday.interfaces
 
 interface Hours {
-    val hours: Int
-    val days: List<Int>?
+    val hours: Double
+    val days: List<Double>?
 }
