+++ src/main/kotlin/model/SickDay.kt
@@ -27,10 +27,10 @@ class SickDay(
     override val from: LocalDate = LocalDate.now(),
     override val to: LocalDate = LocalDate.now(),
 
-    override val hours: Int,
+    override val hours: Double,
 
     @ElementCollection
-    override val days: List<Int>? = null,
+    override val days: List<Double>? = null,
 
     val description: String? = null,
