+++ src/main/kotlin/model/WorkDay.kt
@@ -24,10 +24,10 @@ class WorkDay(
     override val from: LocalDate = LocalDate.now(),
     override val to: LocalDate = LocalDate.now(),
 
-    override val hours: Int,
+    override val hours: Double,
 
     @ElementCollection
-    override val days: List<Int>? = null,
+    override val days: List<Double>? = null,
 
     @ManyToOne
     val assignment: Assignment,
